differentiate between autopilot and uav type - show both in About tab. Refactor this enum to string crap to its own file
