+++ qml/ui/configpopup/AboutPanel.qml
@@ -249,7 +249,7 @@ Rectangle {
         // ------------
         Card {
             id: flightControllBox
-            height: 128
+            height: 150
             Layout.fillWidth: true
             cardName: qsTr("FC")
             cardBody:
@@ -257,6 +257,25 @@ Rectangle {
                 // from https://doc.qt.io/qt-6/qml-qtquick-layouts-rowlayout.html
                 anchors.fill: parent
                 spacing: 2
+                RowLayout{
+                    Layout.fillWidth: true
+                    Layout.minimumHeight: text_minHeight
+                    spacing: 6
+                    Text {
+                        text: qsTr("Autopilot:")
+                        height: 24
+                        font.pixelSize: 14
+                        font.bold: true
+                        leftPadding: 12
+                    }
+                    Text {
+                        text: _fcMavlinkSystem.autopilot_type
+                        height: 24
+                        width: 256
+                        font.pixelSize: 14
+                        leftPadding: 6
+                    }
+                }
                 RowLayout{
                     Layout.fillWidth: true
                     Layout.minimumHeight: text_minHeight
